requirements.txt
Flask==2.0.1
pandas==1.4.3
psycopg2_binary==2.9.3
requests==2.25.1
SQLAlchemy==1.4.40
